requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
fastapi==0.63.0
h11==0.12.0
idna==2.10
numpy==1.19.5
opencv-python==4.5.1.48
Pillow==8.1.0
pydantic==1.7.3
python-multipart==0.0.5
requests==2.25.1
six==1.15.0
starlette==0.13.6
torch==1.7.1
torchvision==0.8.2
tqdm==4.55.1
typing-extensions==3.7.4.3
urllib3==1.26.2
uvicorn==0.13.3
